app/connectors/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic>=2.0.0
prometheus-client
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
